+++ hooks.php
@@ -14,7 +14,7 @@
  * this file to be picked up.
  *
  * @package cwp7
- * @version 2.0.2
+ * @version 2.0.3
  * @author  Booysen Logistics <bradley@booysenlogistics.co.za>
  * @license MIT
  * @link    https://github.com/bradleygb/whmcs-cwp-module